backend/requirements.txt
django==5.1
djangorestframework==3.17.1
drf-spectacular==0.29.0
gunicorn==25.3.0
python-dotenv==1.2.2
django-cors-headers==4.9.0
whitenoise==6.12.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.3
boto3==1.35.0
Pillow==11.2.1
redis==5.2.1
razorpay==1.4.2